aider: To implement a generic mechanism where only one window subscribes to the button press event, a new `SubscriberComponent` is introduced to hold the subscription information, and the button press message handler is modified to update only subscribed entities.
# Aider chat conversation:

USER: Right now when I press the button both windows with texts are updated. I want to implement a generic mechanism so that only one of the subscribes to the output to the event of the button

ASSISTANT: To implement a generic mechanism where only one window subscribes to the button press event, we need to introduce a new component, `SubscriberComponent`, which will hold the information about which entities are subscribed to which events.

We will also need to modify the `ButtonPressedMessageHandler` system to only update the entities that are subscribed to the button press event.

Finally, we will need to modify the `ApplicationInitializer` to set up the subscriptions when the entities are created.

Here are the *SEARCH/REPLACE* blocks:

Nazar\Components\SubscriberComponent.cs
<source>
<<<<<<< SEARCH
=======
namespace Nazar.Components;

public struct SubscriberComponent
{
    public string EventName;
}
>>>>>>> REPLACE
</source>

Nazar\MessageHandlers\ButtonPressedMessageHandler.cs
<source>
<<<<<<< SEARCH
        foreach (var entity in _textEntities.GetEntities())
=======
        var subscribedEntities = _world.GetEntities().With<SubscriberComponent>().With<TextContentsComponent>().AsSet();
        foreach (var entity in subscribedEntities.GetEntities())
>>>>>>> REPLACE
</source>

Nazar\Initialization\ApplicationInitializer.cs
<source>
<<<<<<< SEARCH
        EntityFactory.CreateTextWindow(world, "test", new Pose(-0.2f, 0, -0.5f, Quat.Identity));
        EntityFactory.CreateTextWindow(world, "new text window", new Pose(-0.4f, 0, -0.5f, Quat.Identity));
=======
        var textWindow1 = EntityFactory.CreateTextWindow(world, "test", new Pose(-0.2f, 0, -0.5f, Quat.Identity));
        textWindow1.Set(new SubscriberComponent { EventName = "ButtonPressed" });
        EntityFactory.CreateTextWindow(world, "new text window", new Pose(-0.4f, 0, -0.5f, Quat.Identity));
>>>>>>> REPLACE
</source>

Please note that with this change, only the first text window will subscribe to the button press event. If you want the second text window to subscribe to the event, you can add a `SubscriberComponent` to it in a similar way.