aider: Modified ButtonPressedMessageHandler and ButtonInteractionSystem to update entities based on the specific button pressed.
# Aider chat conversation:

USER: Both are being changed when I press a button

ASSISTANT: I understand that both text windows are being updated when you press either button. This is because the `ButtonPressedMessageHandler` system is currently updating all entities with a `SubscriberComponent` and a `TextContentsComponent` when a button is pressed.

To fix this, we need to modify the `ButtonPressedMessageHandler` system to only update the entities that are subscribed to the specific button that was pressed. We also need to modify the `ButtonInteractionSystem` to publish a `ButtonPressedMessage` with the `EventName` of the `SubscriberComponent` of the button that was pressed.

Here are the *SEARCH/REPLACE* blocks:

Nazar\Systems\ButtonInteractionSystem.cs
<source>
<<<<<<< SEARCH
            World.Publish(new ButtonPressedMessage { Message = "pressed" });
=======
            var eventName = entity.Get<SubscriberComponent>().EventName;
            World.Publish(new ButtonPressedMessage { Message = "pressed", EventName = eventName });
>>>>>>> REPLACE
</source>

Nazar\MessageHandlers\ButtonPressedMessageHandler.cs
<source>
<<<<<<< SEARCH
        var subscribedEntities = _world.GetEntities().With<SubscriberComponent>().With<TextContentsComponent>().AsSet();
        foreach (var entity in subscribedEntities.GetEntities())
=======
        var subscribedEntities = _world.GetEntities().With<SubscriberComponent>(c => c.EventName == message.EventName).With<TextContentsComponent>().AsSet();
        foreach (ref readonly var entity in subscribedEntities.GetEntities())
>>>>>>> REPLACE
</source>

Nazar\Components\ButtonPressedMessage.cs
<source>
<<<<<<< SEARCH
public struct ButtonPressedMessage
{
    public string Message { get; set; }
}
=======
public struct ButtonPressedMessage
{
    public string Message { get; set; }
    public string EventName { get; set; }
}
>>>>>>> REPLACE
</source>